Move XmlNode out of LewisAndShort
