+++ src/common/lewis_and_short/ls_orths.ts
@@ -1,4 +1,4 @@
-import { XmlNode } from "@/common/lewis_and_short/xml_node";
+import { XmlNode } from "@/common/xml_node";
 import { assert } from "../assert";
 import { BREVES, MACRONS } from "../character_utils";
