+++ src/common/lewis_and_short/ls_parser.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from "fs";
-import { XmlNode } from "./xml_node";
+import { XmlNode } from "@/common/xml_node";
 import { extractEntries, parseEntriesInline } from "./ls_xml_utils";
 
 export function getRaw(path: string): string[] {
