+++ src/common/lewis_and_short/ls_styling.test.ts
@@ -1,4 +1,4 @@
-import { XmlNode } from "./xml_node";
+import { XmlNode } from "@/common/xml_node";
 import {
   substituteAbbreviation,
   attachHoverText,
