+++ src/common/lewis_and_short/ls_styling.ts
@@ -1,5 +1,5 @@
 import { checkPresent } from "../assert";
-import { XmlNode } from "./xml_node";
+import { XmlNode } from "@/common/xml_node";
 
 const START_CHARACTERS = new Set<string>(" ();—-");
