+++ src/common/xml_node.ts
