+++ src/web/client/pages/dictionary_utils.test.tsx
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import { XmlNode } from "@/common/lewis_and_short/xml_node";
+import { XmlNode } from "@/common/xml_node";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import user from "@testing-library/user-event";
